project/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0.post2
bcrypt==4.0.1
python-dotenv==1.0.0
